+++ models/cities-in-hong-kong/raw/training/convert-to-wav.sh
@@ -1,0 +1,20 @@
+#!/usr/bin/env bash
+
+ffmpeg -i central-and-western-district.m4a -ar 16000 -ac 1 central-and-western-district.wav
+ffmpeg -i eastern-district.m4a -ar 16000 -ac 1 eastern-district.wav
+ffmpeg -i islands-district.m4a -ar 16000 -ac 1 islands-district.wav
+ffmpeg -i kowloon-city-district.m4a -ar 16000 -ac 1 kowloon-city-district.wav
+ffmpeg -i kwai-tsing-district.m4a -ar 16000 -ac 1 kwai-tsing-district.wav
+ffmpeg -i kwun-tong-district.m4a -ar 16000 -ac 1 kwun-tong-district.wav
+ffmpeg -i north-district.m4a -ar 16000 -ac 1 north-district.wav
+ffmpeg -i sai-kung-district.m4a -ar 16000 -ac 1 sai-kung-district.wav
+ffmpeg -i sha-tin-district.m4a -ar 16000 -ac 1 sha-tin-district.wav
+ffmpeg -i sham-shui-po-district.m4a -ar 16000 -ac 1 sham-shui-po-district.wav
+ffmpeg -i southern-district.m4a -ar 16000 -ac 1 southern-district.wav
+ffmpeg -i tai-po-district.m4a -ar 16000 -ac 1 tai-po-district.wav
+ffmpeg -i tsuen-wan-district.m4a -ar 16000 -ac 1 tsuen-wan-district.wav
+ffmpeg -i tuen-mun-district.m4a -ar 16000 -ac 1 tuen-mun-district.wav
+ffmpeg -i wan-chai-district.m4a -ar 16000 -ac 1 wan-chai-district.wav
+ffmpeg -i wong-tai-sin-district.m4a -ar 16000 -ac 1 wong-tai-sin-district.wav
+ffmpeg -i yau-tsim-mong-district.m4a -ar 16000 -ac 1 yau-tsim-mong-district.wav
+ffmpeg -i yuen-long-district.m4a -ar 16000 -ac 1 yuen-long-district.wav