+++ packages/component/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  setupFilesAfterEnv: [
+    './utils/setupDotEnv.js',
+    './utils/setupGlobalAgent.js',
+    './utils/setupFetchPolyfill.js'
+  ],
+  testEnvironment: 'node'
+};