+++ packages/component/src/SpeechServices/patchOptions.credentials.spec.js
@@ -138,3 +138,86 @@ test('should throw exception for subscriptionKey as () => Promise<number>', asyn
 test('should throw exception when authorizationToken and subscriptionKey are not specified', async () => {
   expect(() => patchOptions({ region: 'westus2' })).toThrow();
 });
+
+test('should throw exception when region and speechSynthesisHostname are not specified', async () => {
+  await expect(
+    patchOptions({ credentials: { authorizationToken: 'AUTHORIZATION_TOKEN' } }).fetchCredentials()
+  ).rejects.toThrow();
+});
+
+test('should throw exception both region and customVoiceHostname are specified', async () => {
+  await expect(
+    patchOptions({
+      credentials: {
+        authorizationToken: 'AUTHORIZATION_TOKEN',
+        customVoiceHostname: 'westus2.cris.ai'
+      }
+    }).fetchCredentials()
+  ).rejects.toThrow();
+});
+
+test('should throw exception both region and speechSynthesisHostname are specified', async () => {
+  await expect(
+    patchOptions({
+      credentials: {
+        authorizationToken: 'AUTHORIZATION_TOKEN',
+        speechSynthesisHostname: 'westus2.tts.speech.microsoft.com'
+      }
+    }).fetchCredentials()
+  ).rejects.toThrow();
+});
+
+test('should throw exception both region and speechRecognitionHostname are specified', async () => {
+  await expect(
+    patchOptions({
+      credentials: {
+        authorizationToken: 'AUTHORIZATION_TOKEN',
+        speechSynthesisHostname: 'westus2.stt.speech.microsoft.com'
+      }
+    }).fetchCredentials()
+  ).rejects.toThrow();
+});
+
+test('should throw exception if only speechRecognitionHostname are specified', async () => {
+  await expect(
+    patchOptions({
+      credentials: {
+        speechRecognitionHostname: 'westus2.stt.speech.microsoft.com'
+      }
+    }).fetchCredentials()
+  ).rejects.toThrow();
+});
+
+test('using custom hostname without Custom Voice', async () => {
+  await expect(
+    patchOptions({
+      credentials: {
+        authorizationToken: 'AUTHORIZATION_TOKEN',
+        speechRecognitionHostname: 'westus2.stt.speech.microsoft.com',
+        speechSynthesisHostname: 'westus2.stt.speech.microsoft.com'
+      }
+    }).fetchCredentials()
+  ).resolves.toEqual({
+    authorizationToken: 'AUTHORIZATION_TOKEN',
+    speechRecognitionHostname: 'westus2.stt.speech.microsoft.com',
+    speechSynthesisHostname: 'westus2.stt.speech.microsoft.com'
+  });
+});
+
+test('using custom hostname with Custom Voice', async () => {
+  await expect(
+    patchOptions({
+      credentials: {
+        authorizationToken: 'AUTHORIZATION_TOKEN',
+        customVoiceHostname: 'westus2.cris.ai',
+        speechRecognitionHostname: 'westus2.stt.speech.microsoft.com',
+        speechSynthesisHostname: 'westus2.stt.speech.microsoft.com'
+      }
+    }).fetchCredentials()
+  ).resolves.toEqual({
+    authorizationToken: 'AUTHORIZATION_TOKEN',
+    customVoiceHostname: 'westus2.cris.ai',
+    speechRecognitionHostname: 'westus2.stt.speech.microsoft.com',
+    speechSynthesisHostname: 'westus2.stt.speech.microsoft.com'
+  });
+});