+++ packages/component/utils/setupFetchPolyfill.js
@@ -1,0 +1,3 @@
+global.fetch = jest.fn(require('node-fetch'));
+
+afterEach(() => global.fetch.mockClear());