+++ CHANGELOG.md
@@ -9,10 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Breaking changes
 
 - As Bing Speech has deprecated and removed from Azure. Bing Speech support has been removed.
+- `fetchAuthorizationToken` is removed, please refer to [this documentation on how to get an access token](https://docs.microsoft.com/en-us/azure/cognitive-services/speech-service/rest-speech-to-text#how-to-get-an-access-token).
+
+### Added
+
+- Resolves [#88](https://github.com/compulim/web-speech-cognitive-services/issues/88). Support sovereign cloud, by [@compulim](https://github.com/compulim) in PR [#89](https://github.com/compulim/web-speech-cognitive-services/pull/89)
 
 ### Removed
 
-- Bing Speech support is removed, in PR [#87](https://github.com/compulim/web-speech-cognitive-services/pull/87), by [@compulim](https://github.com/compulim)
+- Bing Speech support is removed, by [@compulim](https://github.com/compulim), in PR [#87](https://github.com/compulim/web-speech-cognitive-services/pull/87).
+- `fetchAuthorizationToken` is removed, by [@compulim](https://github.com/compulim), in PR [#89](https://github.com/compulim/web-speech-cognitive-services/pull/89).
 
 ### Changed
 
@@ -39,6 +45,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
       - [`lolex@5.1.2`](https://npmjs.com/package/lolex)
       - [`webpack-cli@3.3.11`](https://npmjs.com/package/webpack-cli)
       - [`webpack@4.42.0`](https://npmjs.com/package/webpack)
+- Use [`p-defer-es5`](https://npmjs.com/package/p-defer-es5) instead of adopted `createDeferred.js`, in PR [#89](https://github.com/compulim/web-speech-cognitive-services/pull/89).
 
 ## [6.0.0] - 2019-12-03